fix : SellNowService -> SellNowProvider 로 renaming
