+++ src/test/java/bc1/gream/domain/sell/service/SellNowProviderIntegrationTest.java
@@ -6,6 +6,7 @@ import bc1.gream.domain.buy.entity.Buy;
 import bc1.gream.domain.buy.repository.BuyRepository;
 import bc1.gream.domain.sell.dto.request.SellNowRequestDto;
 import bc1.gream.domain.sell.dto.response.SellNowResponseDto;
+import bc1.gream.domain.sell.provider.SellNowProvider;
 import bc1.gream.test.BaseIntegrationTest;
 import bc1.gream.test.BuyTest;
 import bc1.gream.test.CouponTest;
@@ -20,10 +21,10 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
 @Disabled("통합테스트는 로컬에서만 실행합니다. 실행 시, SECRET KEY 에 대한 IntelliJ 환경변수를 설정해주어야 합니다.")
-class SellNowServiceIntegrationTest extends BaseIntegrationTest implements BuyTest, CouponTest {
+class SellNowProviderIntegrationTest extends BaseIntegrationTest implements BuyTest, CouponTest {
 
     @Autowired
-    private SellNowService sellNowService;
+    private SellNowProvider sellNowProvider;
     @Autowired
     private BuyRepository buyRepository;
 
@@ -52,7 +53,7 @@ class SellNowServiceIntegrationTest extends BaseIntegrationTest implements BuyTe
         SellNowRequestDto requestDto = new SellNowRequestDto(TEST_BUY_PRICE, TEST_GIFTICON_URL);
 
         // WHEN
-        SellNowResponseDto responseDto = sellNowService.sellNowProduct(savedSeller, requestDto, savedProduct.getId());
+        SellNowResponseDto responseDto = sellNowProvider.sellNowProduct(savedSeller, requestDto, savedProduct.getId());
 
         // THEN
         System.out.println("responseDto = " + responseDto);
@@ -70,7 +71,7 @@ class SellNowServiceIntegrationTest extends BaseIntegrationTest implements BuyTe
         List<SellNowResponseDto> sellNowResponseDtos = new ArrayList<>();
         for (int i = 0; i < threadCount; i++) {
             executorService.submit(() -> {
-                SellNowResponseDto responseDto = sellNowService.sellNowProduct(savedSeller, requestDto, savedProduct.getId());
+                SellNowResponseDto responseDto = sellNowProvider.sellNowProduct(savedSeller, requestDto, savedProduct.getId());
                 sellNowResponseDtos.add(responseDto);
 //                try{
 //                } catch (InterruptedException ex) {
